requirements.txt
numpy
pandas
scikit-learn
streamlit

xgboost ==1.6.1 # downgraded from latest 1.7.3
scikeras
catboost

#pytest-cov
tabulate
termcolor
beautifulsoup4

# Was temporarily removed for heroku slug (but restored since using Streamlit and is required)
seaborn #==0.11.2
tensorflow-cpu

# Newly added
pytest

# Temporarily remove for heroku slug
#jupyter #==1.0.0
#matplotlib #==3.5.2
#matplotlib-inline #==0.1.3
#notebook #==6.4.12

#xxxtensorflow

#html2text #==2020.1.16
#ipykernel #==6.13.1
#regex #==2022.7.9
#requests #==2.28.0
#scipy #==1.8.1
